Refactored pylamda
