+++ lib/types.ts
@@ -1,5 +1,5 @@
 import { ITable } from "aws-cdk-lib/aws-dynamodb";
-import { ILayerVersion } from "aws-cdk-lib/aws-lambda";
+import { ILayerVersion, Runtime } from "aws-cdk-lib/aws-lambda";
 import { RetentionDays } from "aws-cdk-lib/aws-logs";
 
 export type FunctionConfig = {
